scaletempo: don't use a window for offset search
This might seem counter intuitive at first, but we want to change the
sound in total as little as possible, not only the middle part of the
overlap.

This doesn't make a noticable difference with default parameters, but
for smaller strides with larger search (e.g. search=30:stride=20)
the difference becomes very noticable.